+++ fabtools/gvm.py
@@ -9,6 +9,14 @@ This module provides tools for installing `GVM`_ : the Groovy enVironment Manage
 """
 
 from fabric.api import run
+from fabric.contrib.files import sed
+
+from fabtools.require.deb import packages as require_deb_packages
+from fabtools.require.oracle_jdk import installed as java
+from fabtools.require.pkg import packages as require_pkg_packages
+from fabtools.require.rpm import packages as require_rpm_packages
+from fabtools.system import UnsupportedFamily, distrib_family
+
 
 def install(java_version=None):
     """
@@ -22,13 +30,6 @@ def install(java_version=None):
         fabtools.gvm.install()
 
     """
-    from fabtools.require.deb import packages as require_deb_packages
-    from fabtools.require.pkg import packages as require_pkg_packages
-    from fabtools.require.rpm import packages as require_rpm_packages
-    from fabtools.require.oracle_jdk import installed as java
-    from fabtools.system import distrib_family, distrib_codename
-    from fabric.contrib.files import sed
-
     res = run('gvm help', quiet=True)
     if res.failed:
         family = distrib_family()
@@ -40,8 +41,8 @@ def install(java_version=None):
         elif family == 'sun':
             require_pkg_packages(packages)
         else:
-            raise NotImplementedError()
-        
+            raise UnsupportedFamily(supported=['debian', 'redhat', 'sun'])
+
         if java_version is None:
             java()
         else: