+++ fabtools/require/deb.py
@@ -218,33 +218,32 @@ def _to_seconds(var):
 
 def uptodate_index(quiet=True, max_age=86400):
     """
-    Update APT package definitions (``apt-get update``) only
-    if specified time since last update already elapsed.
+    Require an up-to-date package index.
 
-    Configure apt (/etc/apt/apt.conf.d/15update-stamp)
-    to create timestamp (/var/lib/apt/periodic/update-success-stamp)
-    always.
+    This will update the package index (using ``apt-get update``) if the last
+    update occured more than *max_age* ago.
 
-    Example::
+    *max_age* can be specified either as an integer (a value in seconds),
+    or as a dictionary whose keys are units (``seconds``, ``minutes``,
+    ``hours``, ``days``, ``weeks``, ``months``) and values are integers.
+    The default value is 1 hour.
+
+    Examples: ::
 
         from fabtools import require
 
-        # do not update in 1 day
+        # Update index if last time was more than 1 day ago
         require.deb.uptodate_index(max_age={'day': 1})
 
-        # do not update in 1 hour and 30 minutes
+        # Update index if last time was more than 1 hour and 30 minutes ago
         require.deb.uptodate_index(max_age={'hour': 1, 'minutes': 30})
 
     """
 
-    # some distribution doesn't create update-success-stamp by default,
-    # that's why this additional configuration is set:
-    STAMP = '/var/lib/apt/periodic/update-success-stamp'
-    CONFIG_PATH = '/etc/apt/apt.conf.d/15update-stamp-fabtools'
-    contents = 'APT::Update::Post-Invoke-Success {"touch %s 2>/dev/null || true";};\n' % STAMP
-    files.file(path=CONFIG_PATH, contents=contents, use_sudo=True)
+    from fabtools.require import file as require_file
+    require_file('/etc/apt/apt.conf.d/15fabtools-update-stamp', contents='''\
+APT::Update::Post-Invoke-Success {"touch /var/lib/apt/periodic/fabtools-update-success-stamp 2>/dev/null || true";};
+''', use_sudo=True)
 
-    # check time, and update index if necessary
-    last = last_update_time()
-    if last is None or system.time() - last > _to_seconds(max_age):
+    if system.time() - last_update_time() > _to_seconds(max_age):
         update_index(quiet=quiet)