+++ fabtools/require/python.py
@@ -2,8 +2,8 @@
 Python environments and packages
 ================================
 
-This module includes tools for using `virtual environments`_
-and installing packages using `pip`_.
+This module provides high-level tools for using Python `virtual environments`_
+and installing Python packages using the `pip`_ installer.
 
 .. _virtual environments: http://www.virtualenv.org/
 .. _pip: http://www.pip-installer.org/
@@ -23,11 +23,11 @@ from fabtools.python_setuptools import (
     install_setuptools,
     is_setuptools_installed,
 )
-from fabtools.system import distrib_family
+from fabtools.system import UnsupportedFamily, distrib_family
 
 
 MIN_SETUPTOOLS_VERSION = '0.7'
-MIN_PIP_VERSION = '1.3.1'
+MIN_PIP_VERSION = '1.5'
 
 
 def setuptools(version=MIN_SETUPTOOLS_VERSION, python_cmd='python'):
@@ -40,24 +40,18 @@ def setuptools(version=MIN_SETUPTOOLS_VERSION, python_cmd='python'):
     .. _setuptools: http://pythonhosted.org/setuptools/
     """
 
-    from fabtools.require.deb import packages as require_deb_packages
-    from fabtools.require.rpm import packages as require_rpm_packages
+    from fabtools.require.deb import package as require_deb_package
+    from fabtools.require.rpm import package as require_rpm_package
 
     if not is_setuptools_installed(python_cmd=python_cmd):
         family = distrib_family()
 
         if family == 'debian':
-            require_deb_packages([
-                'curl',
-                'python-dev',
-            ])
-
+            require_deb_package('python-dev')
         elif family == 'redhat':
-
-            require_rpm_packages([
-                'curl',
-                'python-devel',
-            ])
+            require_rpm_package('python-devel')
+        else:
+            raise UnsupportedFamily(supported=['debian', 'redhat'])
 
         install_setuptools(python_cmd=python_cmd)
 
@@ -76,7 +70,8 @@ def pip(version=MIN_PIP_VERSION, pip_cmd='pip', python_cmd='python'):
         install_pip(python_cmd=python_cmd)
 
 
-def package(pkg_name, url=None, pip_cmd='pip', python_cmd='python', **kwargs):
+def package(pkg_name, url=None, pip_cmd='pip', python_cmd='python',
+            allow_external=False, allow_unverified=False, **kwargs):
     """
     Require a Python package.
 
@@ -85,12 +80,17 @@ def package(pkg_name, url=None, pip_cmd='pip', python_cmd='python', **kwargs):
 
     Package names are case insensitive.
 
+    Starting with version 1.5, pip no longer scrapes insecure external
+    urls by default and no longer installs externally hosted files by
+    default. Use ``allow_external=True`` or ``allow_unverified=True``
+    to change these behaviours.
+
     ::
 
         from fabtools.python import virtualenv
         from fabtools import require
 
-        # Install package system-wide
+        # Install package system-wide (not recommended)
         require.python.package('foo', use_sudo=True)
 
         # Install package in an existing virtual environment
@@ -101,29 +101,68 @@ def package(pkg_name, url=None, pip_cmd='pip', python_cmd='python', **kwargs):
     """
     pip(MIN_PIP_VERSION, python_cmd=python_cmd)
     if not is_installed(pkg_name, pip_cmd=pip_cmd):
-        install(url or pkg_name, pip_cmd=pip_cmd, **kwargs)
+        install(url or pkg_name,
+                pip_cmd=pip_cmd,
+                allow_external=[url or pkg_name] if allow_external else [],
+                allow_unverified=[url or pkg_name] if allow_unverified else [],
+                **kwargs)
 
 
-def packages(pkg_list, pip_cmd='pip', python_cmd='python', **kwargs):
+def packages(pkg_list, pip_cmd='pip', python_cmd='python',
+             allow_external=None, allow_unverified=None, **kwargs):
     """
     Require several Python packages.
 
     Package names are case insensitive.
+
+    Starting with version 1.5, pip no longer scrapes insecure external
+    urls by default and no longer installs externally hosted files by
+    default. Use ``allow_external=['foo', 'bar']`` or
+    ``allow_unverified=['bar', 'baz']`` to change these behaviours
+    for specific packages.
     """
+    if allow_external is None:
+        allow_external = []
+
+    if allow_unverified is None:
+        allow_unverified = []
+
     pip(MIN_PIP_VERSION, python_cmd=python_cmd)
+
     pkg_list = [pkg for pkg in pkg_list if not is_installed(pkg, pip_cmd=pip_cmd)]
     if pkg_list:
-        install(pkg_list, pip_cmd=pip_cmd, **kwargs)
+        install(pkg_list,
+                pip_cmd=pip_cmd,
+                allow_external=allow_external,
+                allow_unverified=allow_unverified,
+                **kwargs)
 
 
-def requirements(filename, pip_cmd='pip', python_cmd='python', **kwargs):
+def requirements(filename, pip_cmd='pip', python_cmd='python',
+                 allow_external=None, allow_unverified=None,  **kwargs):
     """
     Require Python packages from a pip `requirements file`_.
+
+    Starting with version 1.5, pip no longer scrapes insecure external
+    urls by default and no longer installs externally hosted files by
+    default. Use ``allow_external=['foo', 'bar']`` or
+    ``allow_unverified=['bar', 'baz']`` to change these behaviours
+    for specific packages.
+
+    ::
+
+        from fabtools.python import virtualenv
+        from fabtools import require
+
+        # Install requirements in an existing virtual environment
+        with virtualenv('/path/to/venv'):
+            require.python.requirements('requirements.txt')
 
     .. _requirements file: http://www.pip-installer.org/en/latest/requirements.html
     """
     pip(MIN_PIP_VERSION, python_cmd=python_cmd)
-    install_requirements(filename, pip_cmd=pip_cmd, **kwargs)
+    install_requirements(filename, pip_cmd=pip_cmd, allow_external=allow_external,
+                         allow_unverified=allow_unverified, **kwargs)
 
 
 def virtualenv(directory, system_site_packages=False, venv_python=None,