+++ fabtools/require/supervisor.py
@@ -12,7 +12,7 @@ from __future__ import with_statement
 
 from fabtools.files import watch
 from fabtools.supervisor import update_config, process_status, start_process
-from fabtools.system import distrib_family, distrib_id
+from fabtools.system import UnsupportedFamily, distrib_family, distrib_id
 
 
 def process(name, **kwargs):
@@ -83,8 +83,11 @@ def process(name, **kwargs):
     # Upload config file
     if family == 'debian':
         filename = '/etc/supervisor/conf.d/%(name)s.conf' % locals()
-    elif family == 'redhat' or distrib_id() is 'Archlinux':
+    elif family in ['redhat', 'arch']:
         filename = '/etc/supervisord.d/%(name)s.ini' % locals()
+    else:
+        raise UnsupportedFamily(supported=['debian', 'redhat', 'arch'])
+
     with watch(filename, callback=update_config, use_sudo=True):
         require_file(filename, contents='\n'.join(lines), use_sudo=True)